+++ internal/pkg/runtime/engine/oci/create_linux.go
@@ -2,7 +2,7 @@
 //   Apptainer a Series of LF Projects LLC.
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
-// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2025, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -24,6 +24,7 @@ import (
 
 	"github.com/apptainer/apptainer/internal/pkg/cgroups"
 	"github.com/apptainer/apptainer/internal/pkg/instance"
+	"github.com/apptainer/apptainer/internal/pkg/runtime/engine/oci/cdi"
 	"github.com/apptainer/apptainer/internal/pkg/runtime/engine/oci/rpc/client"
 	"github.com/apptainer/apptainer/internal/pkg/util/fs"
 	"github.com/apptainer/apptainer/internal/pkg/util/fs/mount"
@@ -179,6 +180,13 @@ func (e *EngineOperations) CreateContainer(_ context.Context, pid int, rpcConn n
 
 	if err := e.createState(pid); err != nil {
 		return err
+	}
+
+	// Handle CDI device injection if devices were specified
+	if len(e.EngineConfig.Devices) > 0 {
+		if err := cdi.AddCdiDevices(&e.EngineConfig.OciConfig.Spec, e.EngineConfig.Devices, e.EngineConfig.CdiDirs); err != nil {
+			return fmt.Errorf("while setting up CDI devices: %w", err)
+		}
 	}
 
 	rootfs := e.EngineConfig.OciConfig.Root.Path